+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [3.1.0] - 2017-04-21
+### Added
+- Cli command for requeuing tasks. 
+
+### Fixed
+- Being unable to run `queue:maintenance:wait` before db migrations.
+
 ## [3.0.0] - 2017-04-20
 ### Added
 - Queue maintenance mode based on cache solution instead of the filesystem.
@@ -28,7 +35,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - [Breaking change] Use exchange-less way of using producers and consumers
 
-[Unreleased]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/3.0.0...HEAD
+[Unreleased]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/3.1.0...HEAD
+[3.0.0]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/2.3.0...3.0.0
 [2.3.0]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/printedcom/rabbitmq-queue-bundle/compare/2.1.0...2.2.0